Moved softplus activation to it's own file.
