+++ src/activation/mod.rs
@@ -1,12 +1,12 @@
-use std::f64;
-
 pub mod sigmoid;
 pub mod identity;
 pub mod hyperbolictangent;
+pub mod softplus;
 
 pub use self::sigmoid::Sigmoid;
 pub use self::identity::Identity;
 pub use self::hyperbolictangent::HyperbolicTangent;
+pub use self::softplus::SoftPlus;
 
 /// Activation functions
 pub trait Activation {
@@ -15,24 +15,6 @@ pub trait Activation {
     fn calc(&self, x: f64) -> f64;
     // Derivative
     fn derivative(&self, x: f64) -> f64;
-}
-
-pub struct SoftPlus;
-
-impl Activation for SoftPlus {
-    fn new() -> SoftPlus {
-        return SoftPlus;
-    }
-
-    /// Calculates the SoftPlus of input `x`
-    fn calc(&self, x: f64) -> f64 {
-        (1f64 + x.exp()).ln()
-    }
-
-    /// Calculates the Derivative SoftPlus of input `x`
-    fn derivative(&self, x: f64) -> f64 {
-        1f64 / (1f64 + (-x).exp())
-    }
 }
 
 pub struct RectifiedLinearUnit;
@@ -64,20 +46,7 @@ impl Activation for RectifiedLinearUnit {
 #[cfg(test)]
 mod tests {
     use super::Activation;
-    use super::SoftPlus;
     use super::RectifiedLinearUnit;
-
-    #[test]
-    fn softplus_test() {
-        let activation = SoftPlus::new();
-        assert_approx_eq!(activation.calc(-1f64), 0.3132616875f64);
-    }
-
-    #[test]
-    fn softplus_derivative_test() {
-        let activation = SoftPlus::new();
-        assert_approx_eq!(activation.derivative(-1f64), 0.2689414214f64);
-    }
 
     #[test]
     fn rectifiedlinearunit_test() {
